master

+++ helper/src/main/java/com/vsg/helper/common/util/dao/IDaoAllSimpleListRelation.kt
@@ -1,5 +1,5 @@
 package com.vsg.helper.common.util.dao
 
 interface IDaoAllSimpleListRelation<TEntity> {
-    fun viewAllSimpleList(idRelation: Long): List<TEntity>?
+    fun viewAllSimpleList(idRelation: Int): List<TEntity>?
 }
