+++ helper/src/main/java/com/vsg/helper/common/util/dao/IDaoAllTextSearchRelation.kt
@@ -3,5 +3,5 @@ package com.vsg.helper.common.util.dao
 import androidx.lifecycle.LiveData
 
 interface IDaoAllTextSearchRelation {
-    fun viewGetAllTextSearch(idRelation: Long): LiveData<List<String>>
+    fun viewGetAllTextSearch(idRelation: Int): LiveData<List<String>>
 }
