+++ helper/src/main/java/com/vsg/helper/common/util/dao/IDaoAllUpdateIsDefault.kt
@@ -1,5 +1,5 @@
 package com.vsg.helper.common.util.dao
 
 interface IDaoAllUpdateIsDefault {
-    fun updateSetAllIsDefault(idEntity: Long, idRelation: Long)
+    fun updateSetAllIsDefault(idEntity: Int, idRelation: Int)
 }
