+++ helper/src/main/java/com/vsg/helper/common/util/dao/IDaoHastItemRelation.kt
@@ -1,5 +1,5 @@
 package com.vsg.helper.common.util.dao
 
 interface IDaoHastItemRelation {
-    fun viewHasItems(idRelation: Long): Boolean
+    fun viewHasItems(idRelation: Int): Boolean
 }
