+++ helper/src/main/java/com/vsg/helper/common/util/dao/IDaoNextCode.kt
@@ -2,5 +2,5 @@ package com.vsg.helper.common.util.dao
 
 
 interface IDaoNextCode {
-    fun viewNextAutoCode(idRelation: Long): Long
+    fun viewNextAutoCode(idRelation: Int): Int
 }
