+++ helper/src/main/java/com/vsg/helper/common/util/dao/IGenericDaoUpdateEnabled.kt
@@ -3,7 +3,6 @@ package com.vsg.helper.common.util.dao
 import com.vsg.helper.common.model.IEntity
 import com.vsg.helper.common.model.IIsEnabled
 
-
 interface IGenericDaoUpdateEnabled<T> where T : IEntity, T : IIsEnabled {
-    fun updateSetEnabled(data: Long)
+    fun updateSetEnabled(data: Int)
 }
